LayoutTests/media/media-source/media-source-real-timestampoffset-expected.txt


EVENT(sourceopen)
EVENT(updateend)
RUN(sourceBuffer.timestampOffset = 10)
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '10') OK
EXPECTED (sourceBuffer.buffered.end(0) == '16') OK
END OF TEST
